+++ packages/web-components/src/components/va-table/va-table-inner/test/va-table.e2e.ts
@@ -1,5 +1,5 @@
 import { newE2EPage } from '@stencil/core/testing';
-import { axeCheck } from '../../../testing/test-helpers';
+import { axeCheck } from '../../../../testing/test-helpers';
 
 const sortSetup = async (opts = { ascending: true }) => {
   const page = await newE2EPage();
