+++ packages/web-components/src/components/va-table/va-table-inner/va-table-inner.tsx
@@ -1,8 +1,8 @@
 import { Component, Element, Host, Prop, State, h } from '@stencil/core';
-import { isNumeric } from '../../utils/utils';
+import { isNumeric } from '../../../utils/utils';
 import ascendingIcon from '../../assets/sort-arrow-up.svg?format=text';
 import descendingIcon from '../../assets/sort-arrow-down.svg?format=text';
-import { quicksort, reverseQuicksort } from '../../utils/dom-sort';
+import { quicksort, reverseQuicksort } from '../../../utils/dom-sort';
 import classnames from 'classnames';
 
 /**
