+++ packages/web-components/src/components/va-table/va-table-inner/va-table-row.css
